refactor: split editor canvas
